Fix for RMI examples from RMI Trail of Sun's Java Tutorial from http://java.sun.com/docs/books/tutorial/rmi/index.html.
RMInputStream loads its classes using java.rmi.server.RMIClassLoader. The class loader should link the classes it loads, otherwise we get crashes.

The rest is a fix for deserialization. Field types are passed as signatures, not class names. That means, it's not such a good idea, to pass them to Class.forName directly. I'm letting the TypeSignature thingy do the work now (like it was in the original Classpath code), but I'm also passing the class loader through, which is supposed to be used to load the classes.