Add new docbook docs.
"Porting Kaffe to a new platform", by Edouard G. Parmelan

The HTML is generated from the Docbook sources using maven
and the maven-sdocbook-plugin.  I haven't tried to get those
tools to work on Kaffe yet -- when we have them working,
we'll add maven and the plugins to the kaffe-extras set of
tools.