(verify2) fixed a couple of compile-time warnings.
(verify3) refactored all the major method calls to use a Verifier structure instead of an ever growing list of parameters.  it makes keeping track of memory allocation much easier.  it also makes it easier to add other data to the structure that is passed around to every method call.

(verify-type) started to add the support for the supertype list structure, including the structure itself and some methods.  at this point it is just skeleton code and is not used anywhere, as i got sidetracked in refactoring verify3.  eventually it'll be used/updated much like the uninit code.